Reorganize package and remove pip requirements
* Separates packages into .txt files
* Remove puppet repo and install puppet from rdo repo
* Keep cache when $USER != jenkins
* Remove apt-get false warning
* Fix dev package removal (e.g. gcc is no longer shipped)
* Add "checkpoint" profiling to image build
* Removes .git and unused documentation
* Fix edeploy {user,group}add replacement by mngids.py
* Add gnocchi user to ids.table

+++ build_image.sh
@@ -35,7 +35,7 @@ wait_for_lock
 put_lock
 
 . ./role_configrc
-[ -n "$DEBUG" ] && set -x
+[ -z "$DEBUG" ] || set -x
 
 if [ ! -z "${1}" ]; then
     ARTIFACTS_DIR=${1}/image_build
@@ -55,16 +55,21 @@ function build_qcow {
 
 function build_cache {
     [ -z "${SKIP_BUILD}" ] || return
-    CACHE_HASH="SF: $(git log --format=oneline -n 1 $CACHE_DEPS)"
+    CACHE_HASH="$(git log --format=oneline -n 1 $CACHE_DEPS)"
     LOCAL_HASH="$(head -n 1 ${CACHE_PATH}.description 2> /dev/null)"
 
-    echo "(STEP1) ${CACHE_HASH}"
+    echo "(STEP1) Cache: ${CACHE_HASH}"
     if [ "${LOCAL_HASH}" == "${CACHE_HASH}" ]; then
         echo "(STEP1) Already built, remove ${CACHE_PATH}.description to force rebuild"
         return
     fi
     echo "(STEP1) The local cache needs update (was: [${LOCAL_HASH}])"
-    sudo rm -Rf ${CACHE_PATH}*
+    if [ "${USER}" == "jenkins" ]; then
+        echo "(STEP1) Removing ${CACHE_PATH}"
+        sudo umount ${CACHE_PATH}/proc &> /dev/null
+        sudo umount ${CACHE_PATH}/dev &> /dev/null
+        sudo rm -Rf ${CACHE_PATH}*
+    fi
     sudo mkdir -p ${CACHE_PATH}
     (
         set -e
@@ -76,7 +81,7 @@ function build_cache {
     if [ "$?" != "0" ]; then
         echo "(STEP1) FAILED"; sudo rm -Rf ${CACHE_PATH}.description; exit 1;
     fi
-    echo "(STEP1) SUCCESS: ${CACHE_HASH}"
+    echo "(STEP1) SUCCESS ${CACHE_PATH} : ${CACHE_HASH}"
 }
 
 function build_image {
@@ -85,7 +90,7 @@ function build_image {
 
     LOCAL_HASH=$(head -n 1 ${IMAGE_PATH}-${SF_VER}.description 2> /dev/null)
 
-    echo "(STEP2) ${IMAGE_HASH}"
+    echo "(STEP2) Image: ${IMAGE_HASH}" | sed 's/||.*//g'
     if [ "${LOCAL_HASH}" == "${IMAGE_HASH}" ]; then
         echo "(STEP2) Already built, remove ${IMAGE_PATH}-${SF_VER}.description to force rebuild"
         return
@@ -102,7 +107,8 @@ function build_image {
     }
 
     # Copy the cache
-    sudo rsync -a --delete "${CACHE_PATH}/" "${IMAGE_PATH}/"
+    echo "(STEP2) rsync -a --delete '${CACHE_PATH}/' '${IMAGE_PATH}/'"
+    time sudo rsync -a --delete "${CACHE_PATH}/" "${IMAGE_PATH}/"
 
     (
         set -e
@@ -118,13 +124,14 @@ function build_image {
     if [ "$?" != "0" ]; then
         echo "(STEP2) FAILED"; sudo rm -Rf ${IMAGE_PATH}-${SF_VER}.description; exit 1;
     fi
-    echo "(STEP2) SUCCESS: ${IMAGE_HASH}"
+    echo "(STEP2) SUCCESS ${IMAGE_PATH} : ${IMAGE_HASH}"
 }
 
 prepare_buildenv
-build_cache
 # Make sure subproject are available
+echo "(STEP0) Fetch subprojects..."
 ./image/fetch_subprojects.sh || exit 1
+build_cache
 build_image
 if [ -n "$BUILD_QCOW" ]; then
     echo "(STEP3) Building qcow, please wait..."
